Changes:
Created character update method (not much in it right now)
Implemented effects of officials spend on fief loyalty and income.
Implemented effects of garrison spend on fief loyalty.
Implemented effects of bailiff skills on fief loyalty and expenses.
Changed calculation of overlord tax to make it independent of effects of unrest/rebellion (not his concern).