Changes:
Form1: added controls to household and meeting place screens to enable NPCs to be moved via specific route.  Added controls to household and meeting place screens to enable NPCs to be added/removed to/from entourage.  Added control to household screen to allow NPCs to be fired.
Character: minor changes to fireNPC method.