Changes:
Form1: continued with methods associated with siege.  Also added method to check for lifting of a siege due to besieging army moving.
Fief: added attribute to hold siegeID of active siege.
Army: added method to determine if an army is besieging a fief.
Character: added method to get the army being led by the character.