Changes:
Form1: continued work on siege UI and associated methods.
Siege: added updateSiege method to process seasonal update.  Also added method to check whether attrition applies to defending forces (i.e. have stockpiled supplies run out).
Army: amended updateArmy to allow for certain conditions unique to armies involved in a siege.  Added a method to retrieve a specified siege.  Added a method to check if an army is a defending garrison is a siege, or an additional defending army in a siege.