Changes:
Globals_Server: added method to add JournalEvent to pastEvents.
Journal: added attribute (areNewItems) to indicate presence of new (unread) items.  Also added method to add new JournalEvent to events, setting areNewItems to true.
JournalEvent: added method to extract data from a JournalEvent and return it in a string.
Form1: added method to check if event is of interest to player.  Added method to add a new JournalEvent and indicate its presence to the player.