Changes:
Form1: continued with methods associated with marriage.
Globals_Server: added attribute to hold priorities for different types of JournalEntry, based on role of receiving player.
Journal: added attribute to hold current alert priority level.
JournalEntry: added methods to ascertain priority level of an entry and whether it is of interest to the player.